+++ driver-core/src/main/java/com/datastax/driver/core/querybuilder/BuiltStatement.java
@@ -16,6 +16,7 @@
 package com.datastax.driver.core.querybuilder;
 
 import java.nio.ByteBuffer;
+import java.util.ArrayList;
 import java.util.List;
 
 import com.datastax.driver.core.ColumnMetadata;
@@ -29,19 +30,26 @@ abstract class BuiltStatement extends RegularStatement {
 
     private final List<ColumnMetadata> partitionKey;
     private final ByteBuffer[] routingKey;
-    protected final String keyspace;
+    final String keyspace;
 
     private boolean dirty;
     private String cache;
-    protected Boolean isCounterOp;
+    private ByteBuffer[] values;
 
-    protected BuiltStatement(String keyspace) {
+    Boolean isCounterOp;
+
+    // Whether the user has inputed bind markers. If that's the case, we never generate values as
+    // it means the user meant for the statement to be prepared and we shouldn't add our own markers.
+    boolean hasBindMarkers;
+    private boolean forceNoValues;
+
+    BuiltStatement(String keyspace) {
         this.partitionKey = null;
         this.routingKey = null;
         this.keyspace = keyspace;
     }
 
-    protected BuiltStatement(TableMetadata tableMetadata) {
+    BuiltStatement(TableMetadata tableMetadata) {
         this.partitionKey = tableMetadata.getPartitionKey();
         this.routingKey = new ByteBuffer[tableMetadata.getPartitionKey().size()];
         this.keyspace = tableMetadata.getKeyspace().getName();
@@ -49,38 +57,66 @@ abstract class BuiltStatement extends RegularStatement {
 
     @Override
     public String getQueryString() {
-        if (dirty || cache == null) {
-            StringBuilder sb = buildQueryString();
-
-            // Use the same test that String#trim() uses to determine
-            // if a character is a whitespace character.
-            int l = sb.length();
-            while (l > 0 && sb.charAt(l - 1) <= ' ')
-                l -= 1;
-            if (l != sb.length())
-                sb.setLength(l);
-
-            if (l == 0 || sb.charAt(l - 1) != ';')
-                sb.append(';');
-
-            cache = sb.toString();
-            dirty = false;
-        }
+        maybeRebuildCache();
         return cache;
     }
 
-    protected abstract StringBuilder buildQueryString();
+    private void maybeRebuildCache() {
+        if (!dirty && cache != null)
+            return;
 
-    protected void setDirty() {
-        dirty = true;
+        StringBuilder sb;
+        values = null;
+
+        if (hasBindMarkers || forceNoValues) {
+            sb = buildQueryString(null);
+        } else {
+            List<ByteBuffer> l = new ArrayList<ByteBuffer>();
+            sb = buildQueryString(l);
+            if (!l.isEmpty())
+                values = l.toArray(new ByteBuffer[l.size()]);
+        }
+
+        maybeAddSemicolon(sb);
+
+        cache = sb.toString();
+        dirty = false;
     }
 
-    protected boolean isCounterOp() {
+    static StringBuilder maybeAddSemicolon(StringBuilder sb) {
+        // Use the same test that String#trim() uses to determine
+        // if a character is a whitespace character.
+        int l = sb.length();
+        while (l > 0 && sb.charAt(l - 1) <= ' ')
+            l -= 1;
+        if (l != sb.length())
+            sb.setLength(l);
+
+        if (l == 0 || sb.charAt(l - 1) != ';')
+            sb.append(';');
+        return sb;
+    }
+
+    abstract StringBuilder buildQueryString(List<ByteBuffer> variables);
+
+    boolean isCounterOp() {
         return isCounterOp == null ? false : isCounterOp;
     }
 
-    protected void setCounterOp(boolean isCounterOp) {
+    void setCounterOp(boolean isCounterOp) {
         this.isCounterOp = isCounterOp;
+    }
+
+    void checkForBindMarkers(Object value) {
+        dirty = true;
+        if (Utils.containsBindMarker(value))
+            hasBindMarkers = true;
+    }
+
+    void checkForBindMarkers(Utils.Appendeable value) {
+        dirty = true;
+        if (value != null && value.containsBindMarker())
+            hasBindMarkers = true;
     }
 
     // TODO: Correctly document the InvalidTypeException
@@ -117,8 +153,42 @@ abstract class BuiltStatement extends RegularStatement {
 
     @Override
     public ByteBuffer[] getValues() {
-        // I suppose we should have a mode where we keep values in binary?!
-        return null;
+        maybeRebuildCache();
+        return values;
+    }
+
+    @Override
+    public String toString() {
+        if (forceNoValues)
+            return getQueryString();
+
+        return maybeAddSemicolon(buildQueryString(null)).toString();
+    }
+
+    /**
+     * Allows to force this builder to not generate values (through its {@code getValues()} method).
+     * <p>
+     * By default and for performance reasons, the query builder will not
+     * serialize all values provided to strings.  This means that the
+     * {@link #getQueryString} may return a query string with bind markers
+     * (where and when is at the discretion of the builder) and {@link #getValues}
+     * will return the binary values for those markers. This method allows to force
+     * the builder to not generate binary values but rather to serialize them
+     * all in the query string. In practice, this means that if you call
+     * {@code setForceNoValues(true)}, you are guarateed that {@code getValues()}
+     * will return {@code null} and that the string returned by {@code getQueryString()}
+     * will contain no other bind markers than the one inputed by the user.
+     * <p>
+     * Note that this method is mainly useful for debugging purpose. In general,
+     * the default behavior should be the correct and most efficient one.
+     *
+     * @param forceNoValues whether or not this builder may generate values.
+     * @return this statement.
+     */
+    public RegularStatement setForceNoValues(boolean forceNoValues) {
+        this.forceNoValues = forceNoValues;
+        this.dirty = true;
+        return this;
     }
 
     // This is a duplicate of the one in SimpleStatement, but I don't want to expose this publicly so...
@@ -149,9 +219,9 @@ abstract class BuiltStatement extends RegularStatement {
      */
     abstract static class ForwardingStatement<T extends BuiltStatement> extends BuiltStatement {
 
-        protected T statement;
+        T statement;
 
-        protected ForwardingStatement(T statement) {
+        ForwardingStatement(T statement) {
             super((String)null);
             this.statement = statement;
         }
@@ -162,8 +232,8 @@ abstract class BuiltStatement extends RegularStatement {
         }
 
         @Override
-        protected StringBuilder buildQueryString() {
-            throw new UnsupportedOperationException();
+        StringBuilder buildQueryString(List<ByteBuffer> values) {
+            return statement.buildQueryString(values);
         }
 
         @Override
@@ -177,12 +247,7 @@ abstract class BuiltStatement extends RegularStatement {
         }
 
         @Override
-        protected void setDirty() {
-            statement.setDirty();
-        }
-
-        @Override
-        protected boolean isCounterOp() {
+        boolean isCounterOp() {
             return statement.isCounterOp();
         }
 
@@ -223,6 +288,26 @@ abstract class BuiltStatement extends RegularStatement {
         @Override
         public RetryPolicy getRetryPolicy() {
             return statement.getRetryPolicy();
+        }
+
+        @Override
+        public ByteBuffer[] getValues() {
+            return statement.getValues();
+        }
+
+        @Override
+        void checkForBindMarkers(Object value) {
+            statement.checkForBindMarkers(value);
+        }
+
+        @Override
+        void checkForBindMarkers(Utils.Appendeable value) {
+            statement.checkForBindMarkers(value);
+        }
+
+        @Override
+        public String toString() {
+            return statement.toString();
         }
     }
 }